+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed crash due to unhandled singleton type formatting under the `luau` feature flag. ([#358](https://github.com/JohnnyMorganz/StyLua/issues/358))
 
 ## [0.12.1] - 2022-02-01
 ### Fixed